Minor issues in code
Fixed 2 typos.

Some remarks:
I've had a problem with the constant OAUTH2_HTTP_FOUND used in your code, since this was not defined in your package. Maybe you could add it. Also it would be great if you could add a composer.js file with a valid namespace definition so that your code could be integrated more easily.